refactor(website): move emote container component
