+++ website/src/routes/emotes/(directory)/(emotes)/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import Button from "$/components/input/button.svelte";
 	import { DialogMode } from "$/components/dialogs/dialog.svelte";
-	import EmoteContainer from "$/components/emote-container.svelte";
+	import EmoteContainer from "$/components/layout/emote-container.svelte";
 	import LayoutButtons from "$/components/emotes/layout-buttons.svelte";
 	import TabLink from "$/components/tab-link.svelte";
 	import { emotesLayout, defaultEmoteSetDialogMode } from "$/lib/stores";
